+++ src/index.ts
@@ -1,6 +1,7 @@
 const { mrFeatureFlags } = window;
 import { Flag } from '../types';
 
+//Appends isEnabled method to global window object.
 mrFeatureFlags.isEnabled = (flag: string) => {
 	const isFlagExist: Flag | undefined = mrFeatureFlags.flags.find(
 		(item: Flag) => {
